pyproject.toml
[tool.poetry]
name = "fastapitutorial"
version = "0.1.0"
description = ""
authors = ["jakha921 <jakha921@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.104.0"}
greenlet = "^3.0.1"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
